ref(sampling): Split the sampling crate into submodules
