+++ relay-sampling/src/utils.rs
@@ -1,0 +1,4 @@
+/// Returns `true` if this value is equal to `Default::default()`.
+pub fn is_default<T: Default + PartialEq>(t: &T) -> bool {
+    *t == T::default()
+}
